Hernoem berichtenlijst naar berichtensessiecache en wijzig /ophalen naar /_ophalen
- Module: services/berichtenlijst → services/berichtensessiecache
- Package: nl.rijksoverheid.moz.berichtenlijst → nl.rijksoverheid.moz.berichtensessiecache
- Classes: BerichtenlijstResource → BerichtensessiecacheResource, idem Service en Test
- OpenAPI spec: berichtenlijst-api.yaml → berichtensessiecache-api.yaml
- Endpoint: /berichten/ophalen → /berichten/_ophalen (ADR sub-resource conventie)
- C4 model, README, CLAUDE.md, publiccode.yml, cache keys en URNs bijgewerkt
- apisToGenerate=Berichten in POM om gegenereerde OphalenApi uit te sluiten
