--- services/berichtenlijst/src/main/kotlin/nl/rijksoverheid/moz/berichtenlijst/berichten/Bericht.kt
@@ -1,13 +1,0 @@
-package nl.rijksoverheid.moz.berichtenlijst.berichten
-
-import java.time.Instant
-import java.util.UUID
-
-data class Bericht(
-    val berichtId: UUID,
-    val afzender: String,
-    val ontvanger: String,
-    val onderwerp: String,
-    val tijdstip: Instant,
-    val magazijnId: String,
-)
