--- services/berichtenlijst/src/main/resources/application.properties
@@ -1,42 +1,0 @@
-# Application
-quarkus.application.name=berichtenlijst
-quarkus.application.version=0.1.0
-
-# HTTP
-quarkus.http.port=8080
-
-# OpenAPI / Swagger UI (ADR: spec MOET op /openapi.json beschikbaar zijn)
-quarkus.smallrye-openapi.path=/openapi.json
-quarkus.swagger-ui.always-include=true
-quarkus.smallrye-openapi.info-title=FBS Berichtenlijst API
-quarkus.smallrye-openapi.info-version=0.1.0
-
-# Redis
-quarkus.redis.hosts=redis://localhost:6379
-quarkus.redis.devservices.enabled=false
-
-# Magazijnen configuratie
-magazijnen.instances.magazijn-a.url=http://localhost:8081
-magazijnen.instances.magazijn-a.naam=Magazijn A
-magazijnen.instances.magazijn-b.url=http://localhost:8082
-magazijnen.instances.magazijn-b.naam=Magazijn B
-
-# Notificatie Service
-notificatie.service.url=http://localhost:8083
-
-# Logboek Dataverwerkingen (AVG Art. 30)
-logboekdataverwerking.enabled=true
-logboekdataverwerking.service-name=berichtenlijst
-logboekdataverwerking.clickhouse.endpoint=http://localhost:8123
-logboekdataverwerking.clickhouse.username=${CLICKHOUSE_USERNAME:ldv}
-logboekdataverwerking.clickhouse.password=${CLICKHOUSE_PASSWORD:ldv}
-logboekdataverwerking.clickhouse.database=ldv_logging
-logboekdataverwerking.clickhouse.table=logboek_dataverwerkingen
-
-# Quarkus bean discovery voor logboekdataverwerking-wrapper
-quarkus.index-dependency.ldv.group-id=nl.mijnoverheidzakelijk.ldv
-quarkus.index-dependency.ldv.artifact-id=logboekdataverwerking-wrapper
-
-# Logging
-quarkus.log.level=INFO
-quarkus.log.category."nl.rijksoverheid.moz".level=DEBUG
