+++ services/berichtensessiecache/src/main/kotlin/nl/rijksoverheid/moz/berichtensessiecache/berichten/Bericht.kt
@@ -1,0 +1,13 @@
+package nl.rijksoverheid.moz.berichtensessiecache.berichten
+
+import java.time.Instant
+import java.util.UUID
+
+data class Bericht(
+    val berichtId: UUID,
+    val afzender: String,
+    val ontvanger: String,
+    val onderwerp: String,
+    val tijdstip: Instant,
+    val magazijnId: String,
+)
