+++ services/berichtensessiecache/src/main/kotlin/nl/rijksoverheid/moz/berichtensessiecache/magazijn/MagazijnBerichtenResponse.kt
@@ -1,0 +1,9 @@
+package nl.rijksoverheid.moz.berichtensessiecache.magazijn
+
+import nl.rijksoverheid.moz.berichtensessiecache.berichten.Bericht
+
+data class MagazijnBerichtenResponse(
+    val berichten: List<Bericht> = emptyList(),
+    val totalElements: Long = 0,
+    val totalPages: Int = 0,
+)
