+++ services/berichtensessiecache/src/main/kotlin/nl/rijksoverheid/moz/berichtensessiecache/magazijn/MagazijnResult.kt
@@ -1,0 +1,20 @@
+package nl.rijksoverheid.moz.berichtensessiecache.magazijn
+
+import nl.rijksoverheid.moz.berichtensessiecache.berichten.Bericht
+
+sealed class MagazijnResult {
+    abstract val magazijnId: String
+    abstract val naam: String?
+
+    data class Success(
+        override val magazijnId: String,
+        override val naam: String?,
+        val berichten: List<Bericht>,
+    ) : MagazijnResult()
+
+    data class Failure(
+        override val magazijnId: String,
+        override val naam: String?,
+        val error: Throwable,
+    ) : MagazijnResult()
+}
