+++ services/berichtensessiecache/src/main/kotlin/nl/rijksoverheid/moz/berichtensessiecache/magazijn/MagazijnenConfig.kt
@@ -1,0 +1,14 @@
+package nl.rijksoverheid.moz.berichtensessiecache.magazijn
+
+import io.smallrye.config.ConfigMapping
+import java.util.Optional
+
+@ConfigMapping(prefix = "magazijnen")
+interface MagazijnenConfig {
+    fun instances(): Map<String, MagazijnInstance>
+
+    interface MagazijnInstance {
+        fun url(): String
+        fun naam(): Optional<String>
+    }
+}
